Choices Wiki Stats Dataset file added via upload
Fan-made Choices Wiki Stats dataset downloaded from https://choices-stories-you-play.fandom.com/wiki/Statistics and copied to Excel to upload to GitHub as a .csv file. The Wiki Stats page also contains datasets for translations of Choices books in other languages and a dataset that organizes the books by "tags," but this one contains only the English books and does not include tags.